First crack at making the shift key the same color
The deal is that the "normal" key used in this theme
is actually semi-transparent. So I took the dark non-
transparent shift key from the theme and replaced the
dark part with a 25% opacity white.

It doesn't quite match up, though. Unfortunately GIMP
is not great at dealing with partially transparent
individual colors, so I'll have to find some other
tool to figure out exactly what shade that is meant
to be.

Or maybe just copy the signal light onto the "normal"
key images?

Hm... or maybe I don't need a signal light at all?