Make the cheat sheet larger on big screens
This will make it display using Medium text
if there's at least 480dp available. And
small text (the default) if there is not.

Still not actually small enough to avoid
it getting cut-off in landscape...

Probably the longer solution is to switch
to scrolling. Or maybe split it into three
screens.